Stable version

+++ src/main/java/ok/dht/kovalenko/dao/dto/MappedPairedFiles.java
@@ -3,11 +3,11 @@ package ok.dht.kovalenko.dao.dto;
 import java.nio.MappedByteBuffer;
 
 public record MappedPairedFiles(MappedByteBuffer dataFile, MappedByteBuffer indexesFile) {
-    public long dataLimit() {
+    public int dataLimit() {
         return dataFile.limit();
     }
 
-    public long indexesLimit() {
+    public int indexesLimit() {
         return indexesFile.limit();
     }
 }
